Using paging library to handle API pagination

+++ app/src/main/java/com/matheusfroes/swapi/data/model/PendingBookmark.kt
@@ -6,5 +6,5 @@ import android.arch.persistence.room.PrimaryKey
 @Entity(tableName = "pending_bookmarks")
 data class PendingBookmark(
         @PrimaryKey
-        val personId: Long
+        val personId: Int
 )
